Datetimes are compared non-exactly
I'm not 100% sure how timestamping incoming data is handled on a live deployment, but if it's timestamped by when it actually arrived and it (for some reason) arrived a few seconds late, it could throw off the consolidator. This commit changes the way we compare timestamps to a more forgiving way - i.e. by checking if they are less than 60 seconds away from each other.